Lots of updates/ features added
Okay! This is a good update. First, I removed all of the routes from
app.js and placed them into a routes folder for better separation/
organization of code. I then and renamed the "app.js" file on the root
to "server.js" so it would prevent the developer confused confused when
editing with the "app.js" located in "/src/js/"

I then added click events to all of the items in the the deck menu when
clicked. You can now delete a deck, add a deck, open the menu, and
lastly close the menu all async :)

Next step is to rename the deck then open the editor.